refactor(core): move interfaces around
Moves the `SimpleChange` and lifecycle hook interfaces so that they can import other symbols from `core`.
