+++ packages/core/src/change_detection/lifecycle_hooks.ts
