+++ packages/core/src/render3/hooks.ts
@@ -17,7 +17,7 @@ import {
   OnChanges,
   OnDestroy,
   OnInit,
-} from '../interface/lifecycle_hooks';
+} from '../change_detection/lifecycle_hooks';
 import {assertDefined, assertEqual, assertNotEqual} from '../util/assert';
 
 import {assertFirstCreatePass} from './assert';
