+++ packages/core/src/render3/standalone_service.ts
@@ -9,7 +9,7 @@ import {ɵɵinject as inject} from '../di/injector_compatibility';
 import {ɵɵdefineInjectable as defineInjectable} from '../di/interface/defs';
 import {internalImportProvidersFrom} from '../di/provider_collection';
 import {EnvironmentInjector} from '../di/r3_injector';
-import {OnDestroy} from '../interface/lifecycle_hooks';
+import {OnDestroy} from '../change_detection/lifecycle_hooks';
 import {ComponentDef} from './interfaces/definition';
 import {createEnvironmentInjector} from './ng_module_ref';
