+++ packages/core/src/render3/view/directives.ts
@@ -8,7 +8,7 @@
 
 import {RuntimeError, RuntimeErrorCode} from '../../errors';
 import {Writable} from '../../interface/type';
-import {DoCheck, OnChanges, OnInit} from '../../interface/lifecycle_hooks';
+import {DoCheck, OnChanges, OnInit} from '../../change_detection/lifecycle_hooks';
 import {
   assertGreaterThan,
   assertGreaterThanOrEqual,
